Update TODO for keyring testing
